+++ service/postService.go
@@ -6,9 +6,9 @@ import (
 
 type postService interface {
 	CreatePost(userId int64, text string, imageURLs []string) (*model.Post, error)
-	LikePost(userId int, postId int) error
-	UnlikePost(userId int, postId int) error
-	CommentPost(userId int, postId int, text string) error
+	LikePost(userId int64, postId int64) error
+	UnlikePost(userId int64, postId int64) error
+	CommentPost(userId int64, postId int64, text string) error
 	GetPosts() ([]model.Post, error)
 }
 
@@ -25,14 +25,14 @@ func (s *service) CreatePost(userId int64, text string, imageURLs []string) (*mo
 	post.ID = postID
 	return post, nil
 }
-func (s *service) LikePost(userId int, postId int) error {
+func (s *service) LikePost(userId int64, postId int64) error {
 	like := &model.Like{
 		UserID: userId,
 		PostID: postId,
 	}
 	return s.repository.LikePost(like)
 }
-func (s *service) UnlikePost(userId int, postId int) error {
+func (s *service) UnlikePost(userId int64, postId int64) error {
 	like := &model.Like{
 		UserID: userId,
 		PostID: postId,
@@ -43,6 +43,6 @@ func (s *service) GetPosts() ([]model.Post, error) {
 	return s.repository.GetPosts()
 }
 
-func (s *service) CommentPost(userId int, postId int, text string) error {
+func (s *service) CommentPost(userId int64, postId int64, text string) error {
 	return nil
 }
